Uploaded something that sort of works

+++ Project_Source_Code/TeamSpecs.RideAlong/TeamSpecs.RideAlong.UserAdministration/AccountModificationService.cs
@@ -24,7 +24,7 @@ namespace TeamSpecs.RideAlong.UserAdministration
         public IResponse ModifyUserProfile(string userName,DateTime dateOfBirth, string secondaryEmail) 
         {
             IResponse response = new Response();
-            var userHash = _hashService.hashUser(userName, _pepper.RetrievePepper("Account Creation"));
+            var userHash = _hashService.hashUser(userName, (int) _pepper.RetrievePepper("Account Creation"));
 
             #region Validiating Arguements
             if (String.IsNullOrWhiteSpace(userName))
